+++ src/features/interactive-cube/hooks/use-arcball-rotation.ts
