+++ src/features/interactive-cube/hooks/use-is-space-pressed.ts
