+++ src/features/interactive-cube/hooks/use-pegatine-textures.ts
